requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.12
pyodbc==5.2.0
sqlalchemy==2.0.36
google-generativeai==0.8.3
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.0.1
httpx==0.27.2
